+++ docs/atlascli/command/atlas-alerts-settings-enable.txt
@@ -1,8 +1,8 @@
-.. _atlas-alerts-disable:
+.. _atlas-alerts-settings-enable:
 
-====================
-atlas alerts disable
-====================
+============================
+atlas alerts settings enable
+============================
 
 .. default-domain:: mongodb
 
@@ -12,14 +12,14 @@ atlas alerts disable
    :depth: 1
    :class: singlecol
 
-Disables one alert for the specified project.
+Enables one alert for the specified project.
 
 Syntax
 ------
 
 .. code-block::
 
-   atlas alerts disable <alertId> [options]
+   atlas alerts settings enable <alertId> [options]
 
 .. Code end marker, please don't delete this comment
 
@@ -53,7 +53,7 @@ Options
    * - -h, --help
      - 
      - false
-     - help for disable
+     - help for enable
    * - -o, --output
      - string
      - false
