+++ docs/mongocli/command/mongocli-atlas-alerts-settings-disable.txt
@@ -1,8 +1,8 @@
-.. _mongocli-ops-manager-alerts-enable:
+.. _mongocli-atlas-alerts-settings-disable:
 
-==================================
-mongocli ops-manager alerts enable
-==================================
+======================================
+mongocli atlas alerts settings disable
+======================================
 
 .. default-domain:: mongodb
 
@@ -12,14 +12,14 @@ mongocli ops-manager alerts enable
    :depth: 1
    :class: singlecol
 
-Enables one alert for the specified project.
+Disables one alert for the specified project.
 
 Syntax
 ------
 
 .. code-block::
 
-   mongocli ops-manager alerts enable <alertId> [options]
+   mongocli atlas alerts settings disable <alertId> [options]
 
 .. Code end marker, please don't delete this comment
 
@@ -53,7 +53,7 @@ Options
    * - -h, --help
      - 
      - false
-     - help for enable
+     - help for disable
    * - -o, --output
      - string
      - false
