+++ docs/mongocli/command/mongocli-atlas-alerts.txt
@@ -51,8 +51,6 @@ Related Commands
 
 * :ref:`mongocli-atlas-alerts-acknowledge` - Acknowledges one alert for the specified project.
 * :ref:`mongocli-atlas-alerts-describe` - Retrieves one alert for the specified project.
-* :ref:`mongocli-atlas-alerts-disable` - Disables one alert for the specified project.
-* :ref:`mongocli-atlas-alerts-enable` - Enables one alert for the specified project.
 * :ref:`mongocli-atlas-alerts-global` - Retrieves all the global alerts for the specified Ops Manager project.
 * :ref:`mongocli-atlas-alerts-list` - Retrieves all alerts for the specified Atlas project.
 * :ref:`mongocli-atlas-alerts-settings` - Manages alerts configuration for your project.
@@ -64,8 +62,6 @@ Related Commands
 
    acknowledge </command/mongocli-atlas-alerts-acknowledge>
    describe </command/mongocli-atlas-alerts-describe>
-   disable </command/mongocli-atlas-alerts-disable>
-   enable </command/mongocli-atlas-alerts-enable>
    global </command/mongocli-atlas-alerts-global>
    list </command/mongocli-atlas-alerts-list>
    settings </command/mongocli-atlas-alerts-settings>
