+++ docs/mongocli/command/mongocli-cloud-manager-alerts-settings-disable.txt
@@ -1,8 +1,8 @@
-.. _mongocli-ops-manager-alerts-disable:
+.. _mongocli-cloud-manager-alerts-settings-disable:
 
-===================================
-mongocli ops-manager alerts disable
-===================================
+==============================================
+mongocli cloud-manager alerts settings disable
+==============================================
 
 .. default-domain:: mongodb
 
@@ -19,7 +19,7 @@ Syntax
 
 .. code-block::
 
-   mongocli ops-manager alerts disable <alertId> [options]
+   mongocli cloud-manager alerts settings disable <alertId> [options]
 
 .. Code end marker, please don't delete this comment
