+++ docs/mongocli/command/mongocli-ops-manager-alerts-settings-enable.txt
@@ -1,8 +1,8 @@
-.. _mongocli-atlas-alerts-disable:
+.. _mongocli-ops-manager-alerts-settings-enable:
 
-=============================
-mongocli atlas alerts disable
-=============================
+===========================================
+mongocli ops-manager alerts settings enable
+===========================================
 
 .. default-domain:: mongodb
 
@@ -12,14 +12,14 @@ mongocli atlas alerts disable
    :depth: 1
    :class: singlecol
 
-Disables one alert for the specified project.
+Enables one alert for the specified project.
 
 Syntax
 ------
 
 .. code-block::
 
-   mongocli atlas alerts disable <alertId> [options]
+   mongocli ops-manager alerts settings enable <alertId> [options]
 
 .. Code end marker, please don't delete this comment
 
@@ -53,7 +53,7 @@ Options
    * - -h, --help
      - 
      - false
-     - help for disable
+     - help for enable
    * - -o, --output
      - string
      - false
