+++ internal/cli/alerts/alerts_test.go
@@ -27,7 +27,7 @@ func TestBuilder(t *testing.T) {
 	test.CmdValidator(
 		t,
 		Builder(),
-		8,
+		6,
 		[]string{},
 	)
 }
